exchange_api/api/serializer.py
from rest_framework import serializers
from .models import Crypto
from .models import Stock

class CryptoSerializer(serializers.ModelSerializer):
    class Meta:
        model = Crypto
        fields = '__all__'

class StockSerializer(serializers.ModelSerializer):
    class Meta:
        model = Stock
        fields = '__all__'