exchange_api/requirements.txt
django==5.0.1
mssql-django==1.4
python-dotenv
djangorestframework
